+++ ext/jaro_winkler/adj_matrix.h
@@ -1,19 +1,22 @@
 #pragma once
+
+#include "stdint.h"
+
 #define ADJ_MATRIX_DEFAULT_LENGTH 958
 #define ADJ_MATRIX_SEED 9527
 
 typedef struct _node{
   struct _node *next;
-  unsigned long long x, y;
+  uint64_t x, y;
 } Node;
 
 typedef struct{
   Node ***table;
-  unsigned int length;
+  uint32_t length;
 } AdjMatrix;
 
-AdjMatrix* adj_matrix_new    (unsigned int length);
-void       adj_matrix_add    (AdjMatrix *matrix, unsigned long long x, unsigned long long y);
-char       adj_matrix_find   (AdjMatrix *matrix, unsigned long long x, unsigned long long y);
+AdjMatrix* adj_matrix_new    (uint32_t length);
+void       adj_matrix_add    (AdjMatrix *matrix, uint64_t x, uint64_t y);
+char       adj_matrix_find   (AdjMatrix *matrix, uint64_t x, uint64_t y);
 void       adj_matrix_free   (AdjMatrix *matrix);
 AdjMatrix* adj_matrix_default();
