+++ ext/jaro_winkler/code.c
@@ -1,7 +1,8 @@
+#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 
-void utf_char_to_code(char *str, unsigned long long *ret_code, int *ret_byte_length){
+void utf_char_to_code(char *str, uint64_t *ret_code, size_t *ret_byte_length){
   unsigned char first_char = str[0];
   if(first_char >= 252) *ret_byte_length = 6;      // 1111110x
   else if(first_char >= 248) *ret_byte_length = 5; // 111110xx
@@ -13,15 +14,15 @@ void utf_char_to_code(char *str, unsigned long long *ret_code, int *ret_byte_len
   memcpy(ret_code, str, *ret_byte_length);
 }
 
-void string_to_codes(char *str, int length, unsigned long long **ret_codes, int *ret_length){
-  unsigned int code;
+void string_to_codes(char *str, size_t length, uint64_t **ret_codes, size_t *ret_length){
+  uint32_t code;
   char byte_length;
 
   *ret_codes = calloc(length, sizeof(long long));
   *ret_length = 0;
 
-  for(int i = 0; i < length;){
-    int byte_length;
+  for(size_t i = 0; i < length;){
+    size_t byte_length;
     utf_char_to_code(&str[i], &(*ret_codes)[*ret_length], &byte_length);
     *ret_length += 1;
     i += byte_length;
